REFACTOR extracted recommender-specific actions to separate file
